Rename /themes/default folder to /themes/defaultTheme
