+++ styleguide/TestWrapper.js
@@ -1,7 +1,7 @@
 // @flow
 
 import * as React from 'react';
-import colors from '../src/themes/default/colors';
+import colors from '../src/themes/defaultTheme/colors';
 import ThemeProvider from '../src/themes/ThemeProvider';
 import type { Props } from './TestWrapperTypes';
